[예진] 좋아요 카운트 업데이트

+++ src/main/java/himedia/slivermate/mappers/SliverPostMapper.java
@@ -23,8 +23,6 @@ public interface SliverPostMapper {
 	
 	Boolean checkIfUserLikedPost(@Param("post_id") Long post_id, @Param("user_id") int user_id);
 	
-	void incrementLikeCount(@Param("post_id") int post_id);
-	void decrementLikeCount(@Param("post_id") int post_id);
-
-
+	void incrementLikeCount(@Param("post_id") Long post_id);
+	void decrementLikeCount(@Param("post_id") Long post_id);
 }
